Sonar check

+++ user-service/src/main/java/rs/edu/raf/banka/user_service/controller/UserController.java
@@ -75,10 +75,13 @@ public class UserController {
     @ApiResponses(value = {@ApiResponse(code = 200, message = "OK", response = User.class)})
     public ResponseEntity<?>deleteUser(@PathVariable long id) {
         Optional<User> user = userService.getUserById(id);
-        if(user.get() == null){ResponseEntity.badRequest().build();}
-        if(!userService.deleteUser(user.get())){
-            return ResponseEntity.badRequest().body("Can't delete admin");
-        }
+        if(userService.getUserById(id).isPresent()){
+            user = userService.getUserById(id);
+            if(!userService.deleteUser(user.get())){
+                return ResponseEntity.badRequest().body("Can't delete admin");
+            }
+        }else {ResponseEntity.badRequest().build();}
+
         return ResponseEntity.ok().body(user.get().getUsername() + " disabled");
     }
 
@@ -86,10 +89,13 @@ public class UserController {
     @ApiOperation("Edit user with specific id,text fields are with existing data, the user can change them")
     @ApiResponses(value = {@ApiResponse(code = 200, message = "OK", response = User.class)})
     public ResponseEntity<?>editUser(@PathVariable long id, @RequestHeader("Authorization") String token, @RequestBody CreateUserForm createUserForm) {
-        if(userService.getUserById(id).isEmpty()){
+        User user = null;
+        if(userService.getUserById(id).isPresent()){
+           user = userService.getUserById(id).get();
+        }else{
             return ResponseEntity.badRequest().build();
         }
-        User user = userService.getUserById(id).get();
+
         if (!userService.hasEditPermissions(user, token))
             return ResponseEntity.badRequest().build();
         //if(user.isRequiresOtp())
